+++ examples/svg.rs
@@ -141,8 +141,8 @@ fn run(
                 let dpi_factor = window.scale_factor();
                 let size = window.inner_size();
 
-                canvas.set_size(size.width as u32, size.height as u32, dpi_factor as f32);
-                canvas.clear_rect(0, 0, size.width as u32, size.height as u32, Color::rgbf(0.3, 0.3, 0.32));
+                canvas.set_size(size.width, size.height, dpi_factor as f32);
+                canvas.clear_rect(0, 0, size.width, size.height, Color::rgbf(0.3, 0.3, 0.32));
 
                 canvas.save();
                 canvas.translate(200.0, 200.0);
